# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PUBG Account Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts/{id}/status:
    patch:
      operationId: updateAccountStatus
      tags: [accounts]
      summary: Update account marketplace status (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, reserved, under_review, hidden]
              required:
                - status
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts (admin sees all, public sees only active)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, reserved, under_review, hidden, sold, installment]
        - in: query
          name: public
          schema:
            type: boolean
        - in: query
          name: listedBy
          description: "admin = only accounts with no sellerId; seller = only accounts with sellerId; sellerId=N for one seller"
          schema:
            type: string
        - in: query
          name: sellerId
          schema:
            type: integer
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create a new account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: public
          schema:
            type: boolean
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete account (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /accounts/{id}/sell:
    post:
      operationId: sellAccount
      tags: [accounts]
      summary: Sell an account (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellAccountBody"
      responses:
        "200":
          description: Account sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Account:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        accountId:
          type: string
        purchasePrice:
          type: ["number", "null"]
        priceForSale:
          type: number
        finalSoldPrice:
          type: ["number", "null"]
        purchaseDate:
          type: ["string", "null"]
        previousOwnerContact:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, reserved, under_review, hidden, sold, installment]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerContact:
          type: ["string", "null"]
        sellerId:
          type: ["integer", "null"]
        sellerName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/AccountLink"
        pendingLinksCount:
          type: integer
        totalPaid:
          type: ["number", "null"]
        remainingAmount:
          type: ["number", "null"]
        nextDueDate:
          type: ["string", "null"]
        overdueCount:
          type: ["integer", "null"]
        dueSoonCount:
          type: ["integer", "null"]
      required:
        - id
        - title
        - accountId
        - priceForSale
        - status
        - createdAt
        - updatedAt
        - links
        - pendingLinksCount
    CreateAccountBody:
      type: object
      properties:
        title:
          type: string
        accountId:
          type: string
        purchasePrice:
          type: number
        priceForSale:
          type: number
        purchaseDate:
          type: ["string", "null"]
        previousOwnerContact:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
      required:
        - title
        - accountId
        - purchasePrice
        - priceForSale
    UpdateAccountBody:
      type: object
      properties:
        title:
          type: string
        accountId:
          type: string
        purchasePrice:
          type: number
        priceForSale:
          type: number
        purchaseDate:
          type: ["string", "null"]
        previousOwnerContact:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, reserved, under_review, hidden, sold, installment]
    SellAccountBody:
      type: object
      properties:
        customerName:
          type: string
        customerContact:
          type: string
        finalSoldPrice:
          type: number
        paymentType:
          type: string
          enum: [full, installment]
        installmentTotal:
          type: ["number", "null"]
      required:
        - customerName
        - customerContact
        - finalSoldPrice
        - paymentType
    AccountLink:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        type:
          type: string
          enum: [twitter, google, facebook, game_center, apple, tiktok, whatsapp, number, mail]
        login:
          type: string
        password:
          type: string
        value:
          type: string
        status:
          type: string
          enum: [old_owner, my_controlled, transferred]
        isPending:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - accountId
        - type
        - login
        - password
        - value
        - status
        - isPending
        - createdAt
...